# Use generics for metric instrument API