@MrAlias 2023-04-10
I could be helpful to users if we deprecate the old types (as type aliases) for a release.