@MadVikingGod 2023-04-11
If #3971 does get merged, will there be a follow-up to genericitize the options?