@MrAlias 2023-04-11
> If #3971 does get merged, will there be a follow-up to genericitize the options?

Yeah that was my plan :+1:

(I'm already buckled in and ready for the merge conflict pain :wink:).