@MrAlias 2023-04-11
Two things from conversations IRL about this PR:

1. The `WithUnit[N]` and `WithDescription[N]` options are not ideal. They are defined on a type that will never apply to their output.
   - An alternative approach could be to keep the instrument options/configuration non-generic.
     - This means `WithUnit` and `WithDescription` will "make sense", but `With{Int,Float}64Callback` will still need to be two functions
     - It might make sense, if this approach is taken, to move the instrument options/configuration to the `otel/sdk/metric` package. That is only place these types are used and it would follow suit with the `Meter` interface that remains non-generic.
2. Adding a generic parameter means if we in the future extend that parameter it will not be a backwards compatible change for SDKs. It will be similar to our approach of adding methods to interfaces but it is not something we currently call out in our versioning and compatibility policy.