@pellared 2023-04-12
> The project is already not buildable with Go < 1.17 (due to newer API use by logr) and cannot be managed with go mod for any Go < 1.18

1. As far as I see `logr` [supports >= Go 1.15](https://github.com/go-logr/logr/blob/master/.github/workflows/tests.yaml#L12) and[ `go.mod`](https://github.com/go-logr/logr/blob/master/go.mod) says `go 1.16`.
2. Going with generics for API would make it impossible to build an API implementation for Go < 1.18. If we don't use generics the API could still be used for unsupported versions of Go. It may be a good scenario for [automatic instrumentation](https://github.com/open-telemetry/opentelemetry-go-instrumentation).



Initially, I was leaning towards having generics in both API and SDK. However, right now I start to hesitate. I got a feeling that when I see generics in SDK the generics just make the things "more concrete" and remove noise in docs and API. While in the API it makes another layer of abstraction which makes it harder for me to reason about the API (e.g. https://github.com/open-telemetry/opentelemetry-go/pull/3973#issuecomment-1504077176).

I am sorry that my comment is so unprecise. I decided to leave some comment at this point rather than not leaving any feedback at all.